Use newer version of chef and enable sp service
There were some compatibility issues between a
bluepill recipe I was using and the version of chef
running. This has been resolved by using a never
version of chef.

The result is that the chef recipe for signpost,
now installs bluepill recipes, which monitor the
signpost server itself, but also the iodine channel.